examples: remove {bus1-ioctl,helper}.h
Using own headers for common code does reduce the number of lines
required in each example, but they are placing a unnecessary burden on
the reader to lookup different files. Duplicating the lines are worth it
IMO, because it becomes easier to reuse a example for something new.

While at it some minor refactoring was made.

Signed-off-by: Alexander Alemayhu <alexander@alemayhu.com>